dashboard hasil test

+++ helpers/aggregate/rdtage.js
@@ -2,7 +2,7 @@ const { searching, filterSplit } = require('./func/filter')
 const { sumBetweenFunc } = require('./func')
 const { dateFilter } = require('../filter/date')
 
-const converts = { $convert:{ input: '$age', to: 'double' } }
+const converts = { $convert:{ input: '$age', to: 'int' } }
 
 const paramGroup = (group, value) => {
   return {
